Revert "feat: create DRF endpoint to get course index context (#33667)"
This reverts commit aaea6e5b1443bf261e89a45bf24397ff647f8ed4.
