+++ cms/djangoapps/contentstore/rest_api/v1/serializers/__init__.py
@@ -4,7 +4,6 @@ Serializers for v1 contentstore API.
 from .course_details import CourseDetailsSerializer
 from .course_rerun import CourseRerunSerializer
 from .course_team import CourseTeamSerializer
-from .course_index import CourseIndexSerializer
 from .grading import CourseGradingModelSerializer, CourseGradingSerializer
 from .home import CourseHomeSerializer
 from .proctoring import (
