--- cms/djangoapps/contentstore/rest_api/v1/serializers/course_index.py
@@ -1,31 +1,0 @@
-"""
-API Serializers for course index
-"""
-
-from rest_framework import serializers
-
-from cms.djangoapps.contentstore.rest_api.serializers.common import ProctoringErrorListSerializer
-
-
-class InitialIndexStateSerializer(serializers.Serializer):
-    """Serializer for initial course index state"""
-    expanded_locators = serializers.ListSerializer(child=serializers.CharField())
-    locator_to_show = serializers.CharField()
-
-
-class CourseIndexSerializer(serializers.Serializer):
-    """Serializer for course index"""
-    course_release_date = serializers.CharField()
-    course_structure = serializers.DictField()
-    deprecated_blocks_info = serializers.DictField()
-    discussions_incontext_feedback_url = serializers.CharField()
-    discussions_incontext_learnmore_url = serializers.CharField()
-    initial_state = InitialIndexStateSerializer()
-    initial_user_clipboard = serializers.DictField()
-    language_code = serializers.CharField()
-    lms_link = serializers.CharField()
-    mfe_proctored_exam_settings_url = serializers.CharField()
-    notification_dismiss_url = serializers.CharField()
-    proctoring_errors = ProctoringErrorListSerializer(many=True)
-    reindex_link = serializers.CharField()
-    rerun_notification_id = serializers.IntegerField()
