+++ cms/djangoapps/contentstore/rest_api/v1/serializers/proctoring.py
@@ -4,7 +4,6 @@ API Serializers for proctoring
 
 from rest_framework import serializers
 
-from cms.djangoapps.contentstore.rest_api.serializers.common import ProctoringErrorListSerializer
 from xmodule.course_block import get_available_providers
 
 
@@ -30,6 +29,26 @@ class ProctoredExamConfigurationSerializer(serializers.Serializer):
     proctored_exam_settings = ProctoredExamSettingsSerializer()
     available_proctoring_providers = serializers.ChoiceField(get_available_providers())
     course_start_date = serializers.DateTimeField()
+
+
+class ProctoringErrorModelSerializer(serializers.Serializer):
+    """
+    Serializer for proctoring error model item.
+    """
+    deprecated = serializers.BooleanField()
+    display_name = serializers.CharField()
+    help = serializers.CharField()
+    hide_on_enabled_publisher = serializers.BooleanField()
+    value = serializers.CharField()
+
+
+class ProctoringErrorListSerializer(serializers.Serializer):
+    """
+    Serializer for proctoring error list.
+    """
+    key = serializers.CharField()
+    message = serializers.CharField()
+    model = ProctoringErrorModelSerializer()
 
 
 class ProctoringErrorsSerializer(serializers.Serializer):
