+++ cms/djangoapps/contentstore/rest_api/v1/urls.py
@@ -7,7 +7,6 @@ from openedx.core.constants import COURSE_ID_PATTERN
 from .views import (
     CourseDetailsView,
     CourseTeamView,
-    CourseIndexView,
     CourseGradingView,
     CourseRerunView,
     CourseSettingsView,
@@ -59,11 +58,6 @@ urlpatterns = [
         fr'^course_settings/{COURSE_ID_PATTERN}$',
         CourseSettingsView.as_view(),
         name="course_settings"
-    ),
-    re_path(
-        fr'^course_index/{COURSE_ID_PATTERN}$',
-        CourseIndexView.as_view(),
-        name="course_index"
     ),
     re_path(
         fr'^course_details/{COURSE_ID_PATTERN}$',
