+++ cms/djangoapps/contentstore/rest_api/v1/views/__init__.py
@@ -2,7 +2,6 @@
 Views for v1 contentstore API.
 """
 from .course_details import CourseDetailsView
-from .course_index import CourseIndexView
 from .course_team import CourseTeamView
 from .course_rerun import CourseRerunView
 from .grading import CourseGradingView
