+++ cms/djangoapps/contentstore/utils.py
@@ -25,8 +25,7 @@ from pytz import UTC
 from xblock.fields import Scope
 
 from cms.djangoapps.contentstore.toggles import exam_setting_view_enabled
-from common.djangoapps.course_action_state.models import CourseRerunUIStateManager, CourseRerunState
-from common.djangoapps.course_action_state.managers import CourseActionStateItemNotFoundError
+from common.djangoapps.course_action_state.models import CourseRerunUIStateManager
 from common.djangoapps.course_modes.models import CourseMode
 from common.djangoapps.edxmako.services import MakoService
 from common.djangoapps.student import auth
@@ -46,8 +45,6 @@ from common.djangoapps.util.milestones_helpers import (
     get_namespace_choices,
     generate_milestone_namespace
 )
-from common.djangoapps.util.date_utils import get_default_time_display
-from common.djangoapps.xblock_django.api import deprecated_xblocks
 from common.djangoapps.xblock_django.user_service import DjangoXBlockUserService
 from openedx.core import toggles as core_toggles
 from openedx.core.djangoapps.credit.api import get_credit_requirements, is_credit_course
@@ -83,9 +80,7 @@ from cms.djangoapps.contentstore.toggles import (
     # use_xpert_translations_component,
 )
 from cms.djangoapps.models.settings.course_grading import CourseGradingModel
-from cms.djangoapps.models.settings.course_metadata import CourseMetadata
 from xmodule.library_tools import LibraryToolsService
-from xmodule.course_block import DEFAULT_START_DATE  # lint-amnesty, pylint: disable=wrong-import-order
 from xmodule.modulestore import ModuleStoreEnum  # lint-amnesty, pylint: disable=wrong-import-order
 from xmodule.modulestore.django import modulestore  # lint-amnesty, pylint: disable=wrong-import-order
 from xmodule.modulestore.exceptions import ItemNotFoundError  # lint-amnesty, pylint: disable=wrong-import-order
@@ -1675,89 +1670,6 @@ def get_course_videos_context(course_block, pagination_conf, course_key=None):
         # Cached state for transcript providers' credentials (org-specific)
         course_video_context['transcript_credentials'] = get_transcript_credentials_state_for_org(course.id.org)
     return course_video_context
-
-
-def get_course_index_context(request, course_key, course_block=None):
-    """
-    Utils is used to get context of course index outline.
-    It is used for both DRF and django views.
-    """
-
-    from cms.djangoapps.contentstore.views.course import (
-        course_outline_initial_state,
-        _course_outline_json,
-        _deprecated_blocks_info,
-    )
-    from openedx.core.djangoapps.content_staging import api as content_staging_api
-
-    if not course_block:
-        with modulestore().bulk_operations(course_key):
-            course_block = modulestore().get_course(course_key)
-
-    lms_link = get_lms_link_for_item(course_block.location)
-    reindex_link = None
-    if settings.FEATURES.get('ENABLE_COURSEWARE_INDEX', False):
-        if GlobalStaff().has_user(request.user):
-            reindex_link = f"/course/{str(course_key)}/search_reindex"
-    sections = course_block.get_children()
-    course_structure = _course_outline_json(request, course_block)
-    locator_to_show = request.GET.get('show', None)
-
-    course_release_date = (
-        get_default_time_display(course_block.start)
-        if course_block.start != DEFAULT_START_DATE
-        else _("Set Date")
-    )
-
-    settings_url = reverse_course_url('settings_handler', course_key)
-
-    try:
-        current_action = CourseRerunState.objects.find_first(course_key=course_key, should_display=True)
-    except (ItemNotFoundError, CourseActionStateItemNotFoundError):
-        current_action = None
-
-    deprecated_block_names = [block.name for block in deprecated_xblocks()]
-    deprecated_blocks_info = _deprecated_blocks_info(course_block, deprecated_block_names)
-
-    frontend_app_publisher_url = configuration_helpers.get_value_for_org(
-        course_block.location.org,
-        'FRONTEND_APP_PUBLISHER_URL',
-        settings.FEATURES.get('FRONTEND_APP_PUBLISHER_URL', False)
-    )
-    # gather any errors in the currently stored proctoring settings.
-    advanced_dict = CourseMetadata.fetch(course_block)
-    proctoring_errors = CourseMetadata.validate_proctoring_settings(course_block, advanced_dict, request.user)
-
-    user_clipboard = content_staging_api.get_user_clipboard_json(request.user.id, request)
-
-    course_index_context = {
-        'language_code': request.LANGUAGE_CODE,
-        'context_course': course_block,
-        'lms_link': lms_link,
-        'sections': sections,
-        'course_structure': course_structure,
-        'initial_state': course_outline_initial_state(locator_to_show, course_structure) if locator_to_show else None,  # lint-amnesty, pylint: disable=line-too-long
-        'initial_user_clipboard': user_clipboard,
-        'rerun_notification_id': current_action.id if current_action else None,
-        'course_release_date': course_release_date,
-        'settings_url': settings_url,
-        'reindex_link': reindex_link,
-        'deprecated_blocks_info': deprecated_blocks_info,
-        'notification_dismiss_url': reverse_course_url(
-            'course_notifications_handler',
-            current_action.course_key,
-            kwargs={
-                'action_state_id': current_action.id,
-            },
-        ) if current_action else None,
-        'frontend_app_publisher_url': frontend_app_publisher_url,
-        'mfe_proctored_exam_settings_url': get_proctored_exam_settings_url(course_block.id),
-        'advance_settings_url': reverse_course_url('advanced_settings_handler', course_block.id),
-        'proctoring_errors': proctoring_errors,
-        'taxonomy_tags_widget_url': get_taxonomy_tags_widget_url(course_block.id),
-    }
-
-    return course_index_context
 
 
 class StudioPermissionsService:
